+++ lib-l6470dmx/src/json/sparkfundmxparams.cpp
@@ -2,7 +2,7 @@
  * @file sparkfundmxparams.cpp
  *
  */
-/* Copyright (C) 2025 by Arjan van Vught mailto:info@gd32-dmx.org
+/* Copyright (C) 2025-2026 by Arjan van Vught mailto:info@gd32-dmx.org
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -39,21 +39,18 @@
 #include "sparkfundmx.h"
 #include "json/json_parsehelper.h"
 #include "common/utils/utils_flags.h"
- #include "firmware/debug/debug_debug.h"
+#include "firmware/debug/debug_debug.h"
 
 using common::store::l6470dmx::sparkfun::Flags;
 
-namespace json
-{
-SparkFunDmxParams::SparkFunDmxParams()
-{
+namespace json {
+SparkFunDmxParams::SparkFunDmxParams() {
     motor_index_ = common::store::l6470dmx::kMaxMotors;
     ConfigStore::Instance().DmxL6470CopySparkFunGlobal(&store_sparkfun);
 }
 
-SparkFunDmxParams::SparkFunDmxParams(uint32_t motor_index) : motor_index_(motor_index)
-{
-    DEBUG_PRINTF("Motor index: %u", motor_index_);
+SparkFunDmxParams::SparkFunDmxParams(uint32_t motor_index) : motor_index_(motor_index) {
+    DEBUG_PRINTF("Motor index: %u", static_cast<unsigned>(motor_index_));
 
     assert(motor_index < common::store::l6470dmx::kMaxMotors);
     strncpy(file_name_, SparkFunDmxParamsConst::kFileNameMotor, sizeof(file_name_));
@@ -62,17 +59,13 @@ SparkFunDmxParams::SparkFunDmxParams(uint32_t motor_index) : motor_index_(motor_
     ConfigStore::Instance().DmxL6470CopySparkFunIndexed(motor_index, &store_sparkfun);
 }
 
-void SparkFunDmxParams::SetPosition(const char* val, uint32_t len)
-{
-    if (len == 1)
-    {
+void SparkFunDmxParams::SetPosition(const char* val, uint32_t len) {
+    if (len == 1) {
         const auto kCh = val[0];
 
-        if (kCh >= '0' && kCh <= '9')
-        {
+        if (kCh >= '0' && kCh <= '9') {
             const auto kPosition = static_cast<uint8_t>(kCh - '0');
-            if (kPosition < common::store::l6470dmx::kMaxMotors)
-            {
+            if (kPosition < common::store::l6470dmx::kMaxMotors) {
                 store_sparkfun.position = kPosition;
                 store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetPosition, true);
                 return;
@@ -83,10 +76,8 @@ void SparkFunDmxParams::SetPosition(const char* val, uint32_t len)
     store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetPosition, false);
 }
 
-void SparkFunDmxParams::SetSpiCs(const char* val, uint32_t len)
-{
-    if ((len == 0) || (len > 3))
-    {
+void SparkFunDmxParams::SetSpiCs(const char* val, uint32_t len) {
+    if ((len == 0) || (len > 3)) {
         store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetSpiCs, false);
         return;
     }
@@ -95,10 +86,8 @@ void SparkFunDmxParams::SetSpiCs(const char* val, uint32_t len)
     store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetSpiCs, true);
 }
 
-void SparkFunDmxParams::SetResetPin(const char* val, uint32_t len)
-{
-    if ((len == 0) || (len > 3))
-    {
+void SparkFunDmxParams::SetResetPin(const char* val, uint32_t len) {
+    if ((len == 0) || (len > 3)) {
         store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetResetPin, false);
         return;
     }
@@ -107,10 +96,8 @@ void SparkFunDmxParams::SetResetPin(const char* val, uint32_t len)
     store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetResetPin, true);
 }
 
-void SparkFunDmxParams::SetBusyPin(const char* val, uint32_t len)
-{
-    if ((len == 0) || (len > 3))
-    {
+void SparkFunDmxParams::SetBusyPin(const char* val, uint32_t len) {
+    if ((len == 0) || (len > 3)) {
         store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetBusyPin, false);
         return;
     }
@@ -119,18 +106,14 @@ void SparkFunDmxParams::SetBusyPin(const char* val, uint32_t len)
     store_sparkfun.flags = common::SetFlagValue(store_sparkfun.flags, Flags::Flag::kIsSetBusyPin, true);
 }
 
-void SparkFunDmxParams::Store(const char* buffer, uint32_t buffer_size)
-{
+void SparkFunDmxParams::Store(const char* buffer, uint32_t buffer_size) {
     store_sparkfun.flags = 0;
 
     ParseJsonWithTable(buffer, buffer_size, kSparkFunKeys);
-    
-    if (motor_index_ >= common::store::l6470dmx::kMaxMotors)
-    {
+
+    if (motor_index_ >= common::store::l6470dmx::kMaxMotors) {
         ConfigStore::Instance().DmxL6470StoreSparkFunGlobal(&store_sparkfun);
-    }
-    else
-    {
+    } else {
         ConfigStore::Instance().DmxL6470StoreSparkFunIndexed(motor_index_, &store_sparkfun);
     }
 #ifndef NDEBUG
@@ -138,43 +121,33 @@ void SparkFunDmxParams::Store(const char* buffer, uint32_t buffer_size)
 #endif
 }
 
-void SparkFunDmxParams::Set(SparkFunDmx* sparkfundmx)
-{
+void SparkFunDmxParams::Set(SparkFunDmx* sparkfundmx) {
     assert(sparkfundmx != nullptr);
     const auto kFlags = store_sparkfun.flags;
 
-    if (motor_index_ >= common::store::l6470dmx::kMaxMotors)
-    {
+    if (motor_index_ >= common::store::l6470dmx::kMaxMotors) {
         DEBUG_PUTS("Set Global");
         // #if !defined(H3)
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs)) {
             sparkfundmx->SetGlobalSpiCs(store_sparkfun.spi_cs);
         }
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin)) {
             sparkfundmx->SetGlobalResetPin(store_sparkfun.reset_pin);
         }
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin)) {
             sparkfundmx->SetGlobalBusyPin(store_sparkfun.busy_pin);
         }
         // #endif
-    }
-    else
-    {
+    } else {
         DEBUG_PUTS("Set Local");
         // #if !defined(H3)
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs)) {
             sparkfundmx->SetLocalSpiCs(store_sparkfun.spi_cs);
         }
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin)) {
             sparkfundmx->SetLocalResetPin(store_sparkfun.reset_pin);
         }
-        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin))
-        {
+        if (common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin)) {
             sparkfundmx->SetLocalBusyPin(store_sparkfun.busy_pin);
         }
         // #endif
@@ -185,32 +158,18 @@ void SparkFunDmxParams::Set(SparkFunDmx* sparkfundmx)
 #endif
 }
 
-void SparkFunDmxParams::Dump()
-{
-    if (motor_index_ >= common::store::l6470dmx::kMaxMotors)
-    {
+void SparkFunDmxParams::Dump() {
+    if (motor_index_ >= common::store::l6470dmx::kMaxMotors) {
         printf("%s::%s \'%s\' (global settings):\n", __FILE__, __FUNCTION__, SparkFunDmxParamsConst::kFileName);
-    }
-    else
-    {
+    } else {
         printf("%s::%s \'%s\' :\n", __FILE__, __FUNCTION__, file_name_);
     }
 
     const auto kFlags = store_sparkfun.flags;
 
-    printf(" %s=%u [%d]\n", SparkFunDmxParamsConst::kPosition.name, store_sparkfun.position,
-           common::IsFlagSet(kFlags, Flags::Flag::kIsSetPosition));
-
-    // #if !defined(H3)
-
-    printf(" %s=%u [%d]\n", SparkFunDmxParamsConst::kSpiCs.name, store_sparkfun.spi_cs,
-           common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs));
-
-    // #endif
-
-    printf(" %s=%u [%d]\n", SparkFunDmxParamsConst::kResetPin.name, store_sparkfun.reset_pin,
-           common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin));
-    printf(" %s=%u [%d]\n", SparkFunDmxParamsConst::kBusyPin.name, store_sparkfun.busy_pin,
-           common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin));
+    printf(" %s=%u [%u]\n", SparkFunDmxParamsConst::kPosition.name, static_cast<unsigned>(store_sparkfun.position), static_cast<unsigned>(common::IsFlagSet(kFlags, Flags::Flag::kIsSetPosition)));
+    printf(" %s=%u [%u]\n", SparkFunDmxParamsConst::kSpiCs.name, static_cast<unsigned>(store_sparkfun.spi_cs), static_cast<unsigned>(common::IsFlagSet(kFlags, Flags::Flag::kIsSetSpiCs)));
+    printf(" %s=%u [%u]\n", SparkFunDmxParamsConst::kResetPin.name, static_cast<unsigned>(store_sparkfun.reset_pin), static_cast<unsigned>(common::IsFlagSet(kFlags, Flags::Flag::kIsSetResetPin)));
+    printf(" %s=%u [%u]\n", SparkFunDmxParamsConst::kBusyPin.name, static_cast<unsigned>(store_sparkfun.busy_pin), static_cast<unsigned>(common::IsFlagSet(kFlags, Flags::Flag::kIsSetBusyPin)));
 }
 } // namespace json
